Encapsulate download args validation in newDownloadContext
This validation should be encapsulated in the constructor so it is
covered when `doctor` creates a downloadContext.

This results in duplicate knowledge in `download` cmd validating flags:

    if uuid != "" && slug != "" || uuid == slug {
        return errors.New("need an --exercise name or a solution --uuid")
    }

To avoid duplication knowledge of validation, we create the
`downloadArgsError` interface to allow a single source of validation
with implementation specific error messages.