Add new downloadCmdContext type
`downloadCmdContext` is a context for the running download cmd.
Previously, `runDownload` had several lines of creation logic to create
a context for the download to do the work we actually care about. To
better separate the what from the how, we encapsulate this is in the new
type. This gives a type to hold behavior for downloadCmd itself (such as
`printResult`) that is easily extensible.